Use generics for tester workflow result
