+++ samples/subworkflow/subworkflow_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func Test_Workflow(t *testing.T) {
-	tester := tester.NewWorkflowTester(ParentWorkflow)
+	tester := tester.NewWorkflowTester[any](ParentWorkflow)
 
 	tester.Registry().RegisterWorkflow(SubWorkflow)
 
@@ -21,9 +21,7 @@ func Test_Workflow(t *testing.T) {
 
 	require.True(t, tester.WorkflowFinished())
 
-	var wr int
-	var werr string
-	tester.WorkflowResult(&wr, &werr)
+	wr, werr := tester.WorkflowResult()
 	require.Empty(t, wr)
 	require.Empty(t, werr)
 	tester.AssertExpectations(t)
