+++ tester/tester_test.go
@@ -17,19 +17,18 @@ func Test_Workflow(t *testing.T) {
 		return 0, nil
 	}
 
-	tester := NewWorkflowTester(workflowWithoutActivity)
+	tester := NewWorkflowTester[int](workflowWithoutActivity)
 
 	tester.Execute()
 
 	require.True(t, tester.WorkflowFinished())
-	var wr int
-	tester.WorkflowResult(&wr, nil)
+	wr, _ := tester.WorkflowResult()
 	require.Equal(t, 0, wr)
 	tester.AssertExpectations(t)
 }
 
 func Test_WorkflowBlocked(t *testing.T) {
-	tester := NewWorkflowTester(workflowBlocked)
+	tester := NewWorkflowTester[any](workflowBlocked)
 
 	require.Panics(t, func() {
 		tester.Execute()
@@ -44,37 +43,34 @@ func workflowBlocked(ctx workflow.Context) error {
 }
 
 func Test_Activity(t *testing.T) {
-	tester := NewWorkflowTester(workflowWithActivity)
+	tester := NewWorkflowTester[int](workflowWithActivity)
 
 	tester.OnActivity(activity1, mock.Anything).Return(42, nil)
 
 	tester.Execute()
 
 	require.True(t, tester.WorkflowFinished())
-	var wr int
-	tester.WorkflowResult(&wr, nil)
+	wr, _ := tester.WorkflowResult()
 	require.Equal(t, 42, wr)
 	tester.AssertExpectations(t)
 }
 
 func Test_FailingActivity(t *testing.T) {
-	tester := NewWorkflowTester(workflowWithActivity)
+	tester := NewWorkflowTester[int](workflowWithActivity)
 
 	tester.OnActivity(activity1, mock.Anything).Return(0, errors.New("error"))
 
 	tester.Execute()
 
 	require.True(t, tester.WorkflowFinished())
-	var wr int
-	var werr string
-	tester.WorkflowResult(&wr, &werr)
+	wr, werr := tester.WorkflowResult()
 	require.Equal(t, 0, wr)
 	require.Equal(t, "error", werr)
 	tester.AssertExpectations(t)
 }
 
 // func Test_InvalidActivityMock(t *testing.T) {
-// 	tester := NewWorkflowTester(workflowWithActivity)
+// 	tester := NewWorkflowTester[int](workflowWithActivity)
 
 // 	tester.OnActivity(activityPanics, mock.Anything).Return(1, 2, 3)
 
@@ -87,7 +83,7 @@ func Test_FailingActivity(t *testing.T) {
 // }
 
 func Test_Activity_Retries(t *testing.T) {
-	tester := NewWorkflowTester(workflowWithActivity)
+	tester := NewWorkflowTester[int](workflowWithActivity)
 
 	// Return two errors
 	tester.OnActivity(activity1, mock.Anything).Return(0, errors.New("error")).Once()
@@ -95,22 +91,19 @@ func Test_Activity_Retries(t *testing.T) {
 
 	tester.Execute()
 
-	var r int
-	tester.WorkflowResult(&r, nil)
+	r, _ := tester.WorkflowResult()
 	require.Equal(t, 42, r)
 }
 
 func Test_Activity_WithoutMock(t *testing.T) {
-	tester := NewWorkflowTester(workflowWithActivity)
+	tester := NewWorkflowTester[int](workflowWithActivity)
 
 	tester.Registry().RegisterActivity(activity1)
 
 	tester.Execute()
 
 	require.True(t, tester.WorkflowFinished())
-	var r int
-	var errStr string
-	tester.WorkflowResult(&r, &errStr)
+	r, errStr := tester.WorkflowResult()
 	require.Zero(t, errStr)
 	require.Equal(t, 23, r)
 	tester.AssertExpectations(t)
@@ -134,7 +127,7 @@ func activity1(ctx context.Context) (int, error) {
 }
 
 func Test_Activity_LongRunning(t *testing.T) {
-	tester := NewWorkflowTester(workflowLongRunningActivity)
+	tester := NewWorkflowTester[any](workflowLongRunningActivity)
 	tester.Registry().RegisterActivity(activityLongRunning)
 
 	tester.Execute()
@@ -155,14 +148,13 @@ func activityLongRunning(ctx context.Context) (int, error) {
 }
 
 func Test_Timer(t *testing.T) {
-	tester := NewWorkflowTester(workflowTimer)
+	tester := NewWorkflowTester[timerResult](workflowTimer)
 	start := tester.Now()
 
 	tester.Execute()
 
 	require.True(t, tester.WorkflowFinished())
-	var wr timerResult
-	tester.WorkflowResult(&wr, nil)
+	wr, _ := tester.WorkflowResult()
 	require.True(t, start.Equal(wr.T1))
 
 	e := start.Add(30 * time.Second)
@@ -190,15 +182,14 @@ func workflowTimer(ctx workflow.Context) (timerResult, error) {
 }
 
 func Test_TimerCancellation(t *testing.T) {
-	tester := NewWorkflowTester(workflowTimerCancellation)
+	tester := NewWorkflowTester[time.Time](workflowTimerCancellation)
 	start := tester.Now()
 
 	tester.Execute()
 
 	require.True(t, tester.WorkflowFinished())
 
-	var wfR time.Time
-	tester.WorkflowResult(&wfR, nil)
+	wfR, _ := tester.WorkflowResult()
 	require.True(t, start.Equal(wfR), "expected %v, got %v", start, wfR)
 }
 
@@ -213,7 +204,7 @@ func workflowTimerCancellation(ctx workflow.Context) (time.Time, error) {
 }
 
 func Test_Signals(t *testing.T) {
-	tester := NewWorkflowTester(workflowSignal)
+	tester := NewWorkflowTester[string](workflowSignal)
 	tester.ScheduleCallback(time.Duration(5*time.Second), func() {
 		tester.SignalWorkflow("signal", "s42")
 	})
@@ -222,8 +213,7 @@ func Test_Signals(t *testing.T) {
 
 	require.True(t, tester.WorkflowFinished())
 
-	var wfR string
-	tester.WorkflowResult(&wfR, nil)
+	wfR, _ := tester.WorkflowResult()
 	require.Equal(t, wfR, "s42")
 	tester.AssertExpectations(t)
 }
